LOOCV for Ridge2 model

LOOCV for Random Vector functional link network model with 2 regularization parameters

loocvridge2f(
  y,
  xreg = NULL,
  h = 5,
  level = 95,
  lags = 1,
  nb_hidden = 5,
  nodes_sim = c("sobol", "halton", "unif"),
  activ = c("relu", "sigmoid", "tanh", "leakyrelu", "elu", "linear"),
  a = 0.01,
  lambda_1 = 0.1,
  lambda_2 = 0.1,
  dropout = 0,
  type_forecast = c("recursive", "direct"),
  type_pi = c("gaussian", "bootstrap", "blockbootstrap", "movingblockbootstrap",
    "rvinecopula", "splitconformal"),
  block_length = NULL,
  margins = c("gaussian", "empirical", "student"),
  seed = 1,
  B = 100L,
  type_aggregation = c("mean", "median"),
  centers = NULL,
  type_clustering = c("kmeans", "hclust"),
  ym = NULL,
  cl = 1L,
  show_progress = TRUE,
  ...
)

Arguments

y

A multivariate time series of class ts (preferred) or a matrix

xreg

External regressors. A data.frame (preferred) or a matrix

h

Forecasting horizon

level

Confidence level for prediction intervals

lags

Number of lags

nb_hidden

Number of nodes in hidden layer

nodes_sim

Type of simulation for nodes in the hidden layer

activ

Activation function

a

Hyperparameter for activation function "leakyrelu", "elu"

lambda_1

Regularization parameter for original predictors

lambda_2

Regularization parameter for transformed predictors

dropout

dropout regularization parameter (dropping nodes in hidden layer)

type_forecast

Recursive or direct forecast

type_pi

Type of prediction interval currently "gaussian", "bootstrap", "blockbootstrap", "movingblockbootstrap", "splitconformal" (very experimental right now), "rvinecopula" (with Gaussian margins for now, Student-t coming soon)

block_length

Length of block for circular or moving block bootstrap

margins

Distribution of margins: "gaussian", "empirical", "student" (postponed or never) for type_pi == "rvinecopula"

seed

Reproducibility seed for random stuff

B

Number of bootstrap replications or number of simulations (yes, 'B' is unfortunate)

type_aggregation

Type of aggregation, ONLY for bootstrapping; either "mean" or "median"

centers

Number of clusters for type_clustering

type_clustering

"kmeans" (K-Means clustering) or "hclust" (Hierarchical clustering)

ym

Univariate time series (stats::ts) of yield to maturities with frequency = frequency(y) and start = tsp(y)[2] + 1 / frequency(y). Default is NULL.

cl

An integer; the number of clusters for parallel execution, for bootstrap

show_progress

A boolean; show progress bar for bootstrapping? Default is TRUE.

...

Additional parameters to be passed to kmeans or hclust

Value

An object of class "mtsforecast"; a list containing the following elements:

method

The name of the forecasting method as a character string

mean

Point forecasts for the time series

lower

Lower bound for prediction interval

upper

Upper bound for prediction interval

sims

Model simulations for bootstrapping (basic, or block)

x

The original time series

residuals

Residuals from the fitted model

coefficients

Regression coefficients for type_pi == 'gaussian' for now

References

Moudiki, T., Planchet, F., & Cousin, A. (2018). Multiple time series forecasting using quasi-randomized functional link neural networks. Risks, 6(1), 22.

Author

T. Moudiki

Examples

require(fpp)

print(ahead::loocvridge2f(fpp::insurance))
#> $mean
#>            Quotes TV.advert
#> May 2005 14.76393  8.947782
#> Jun 2005 14.61841  8.862613
#> Jul 2005 14.56234  8.841193
#> Aug 2005 14.49305  8.809091
#> Sep 2005 14.43100  8.781208
#> 
#> $lower
#>            Quotes TV.advert
#> May 2005 11.96618  7.011446
#> Jun 2005 11.82066  6.926276
#> Jul 2005 11.76459  6.904857
#> Aug 2005 11.69530  6.872755
#> Sep 2005 11.63325  6.844872
#> 
#> $upper
#>            Quotes TV.advert
#> May 2005 17.56168  10.88412
#> Jun 2005 17.41616  10.79895
#> Jul 2005 17.36009  10.77753
#> Aug 2005 17.29080  10.74543
#> Sep 2005 17.22875  10.71754
#> 
#> $sims
#> NULL
#> 
#> $x
#>            Quotes TV.advert
#> Jan 2002 12.97065  7.212725
#> Feb 2002 15.38714  9.443570
#> Mar 2002 13.22957  7.534250
#> Apr 2002 12.97065  7.212725
#> May 2002 15.38714  9.443570
#> Jun 2002 11.72288  6.415215
#> Jul 2002 10.06177  5.806990
#> Aug 2002 10.82279  6.203600
#> Sep 2002 13.28707  7.586430
#> Oct 2002 14.57832  8.004935
#> Nov 2002 15.60542  8.834980
#> Dec 2002 15.93515  8.957255
#> Jan 2003 16.99486  9.532990
#> Feb 2003 16.87821  9.392950
#> Mar 2003 16.45128  8.918560
#> Apr 2003 15.28118  8.374120
#> May 2003 15.88901  9.844505
#> Jun 2003 15.67747  9.849390
#> Jul 2003 13.28780  8.402730
#> Aug 2003 12.64484  7.920675
#> Sep 2003 11.82771  7.436085
#> Oct 2003  9.69184  6.340490
#> Nov 2003 10.30415  6.939995
#> Dec 2003 11.38253  6.977100
#> Jan 2004 12.95149  8.010201
#> Feb 2004 13.63092  9.565460
#> Mar 2004  9.12098  6.272510
#> Apr 2004  8.39468  5.707495
#> May 2004 12.30076  7.963540
#> Jun 2004 13.84831  8.494221
#> Jul 2004 15.96246  9.789085
#> Aug 2004 14.19738  8.692825
#> Sep 2004 12.85922  8.057230
#> Oct 2004 12.08837  7.588995
#> Nov 2004 12.93375  8.244881
#> Dec 2004 11.72235  6.675540
#> Jan 2005 15.47126  9.219604
#> Feb 2005 18.43898 10.963800
#> Mar 2005 17.49186 10.456290
#> Apr 2005 14.49168  8.728600
#> 
#> $level
#> [1] 95
#> 
#> $method
#> [1] "ridge2"
#> 
#> $fitted
#>             Quotes TV.advert
#> Jan 2002 13.983454  8.113042
#> Feb 2002 14.310140  8.680078
#> Mar 2002 14.204085  8.367422
#> Apr 2002 13.983454  8.113042
#> May 2002 14.310140  8.680078
#> Jun 2002 13.136549  7.573500
#> Jul 2002 12.139564  7.269130
#> Aug 2002 12.630260  7.516240
#> Sep 2002 14.247551  8.406368
#> Oct 2002 15.085483  8.610639
#> Nov 2002 15.834616  9.129753
#> Dec 2002 16.037153  9.175375
#> Jan 2003 16.609951  9.409639
#> Feb 2003 16.579674  9.347087
#> Mar 2003 16.452555  9.136445
#> Apr 2003 15.691303  8.874788
#> May 2003 14.054876  8.529575
#> Jun 2003 13.434843  8.229103
#> Jul 2003 12.715293  7.977976
#> Aug 2003 12.998432  8.104914
#> Sep 2003 12.590713  7.879126
#> Oct 2003 10.931512  7.005044
#> Nov 2003 10.416928  6.781919
#> Dec 2003 13.045036  8.079712
#> Jan 2004 13.471769  8.342884
#> Feb 2004  8.822228  6.015318
#> Mar 2004  9.707979  6.398258
#> Apr 2004  9.811491  6.420448
#> May 2004 11.936553  7.582567
#> Jun 2004 13.921151  8.555557
#> Jul 2004 14.494976  8.746341
#> Aug 2004 14.079472  8.618854
#> Sep 2004 13.046936  8.135087
#> Oct 2004 12.726572  7.953926
#> Nov 2004 12.426974  7.834539
#> Dec 2004 13.211185  7.811190
#> Jan 2005 15.485788  9.264229
#> Feb 2005 16.598686  9.685778
#> Mar 2005 16.037908  9.448414
#> 
#> $residuals
#>               Quotes   TV.advert
#> Jan 2002  1.40368643  1.33052782
#> Feb 2002 -1.08056990 -1.14582835
#> Mar 2002 -1.23343504 -1.15469695
#> Apr 2002  1.40368643  1.33052782
#> May 2002 -2.58725990 -2.26486335
#> Jun 2002 -3.07477874 -1.76651025
#> Jul 2002 -1.31677412 -1.06553020
#> Aug 2002  0.65681006  0.07019039
#> Sep 2002  0.33076863 -0.40143292
#> Oct 2002  0.51993723  0.22434076
#> Nov 2002  0.10053432 -0.17249791
#> Dec 2002  0.95770684  0.35761548
#> Jan 2003  0.26825909 -0.01668861
#> Feb 2003 -0.12839403 -0.42852687
#> Mar 2003 -1.17137522 -0.76232506
#> Apr 2003  0.19770654  0.96971721
#> May 2003  1.62259434  1.31981494
#> Jun 2003 -0.14704277  0.17362676
#> Jul 2003 -0.07045343 -0.05730098
#> Aug 2003 -1.17072203 -0.66882907
#> Sep 2003 -2.89887258 -1.53863637
#> Oct 2003 -0.62736155 -0.06504866
#> Nov 2003  0.96560231  0.19518092
#> Dec 2003 -0.09354593 -0.06951135
#> Jan 2004  0.15915097  1.22257588
#> Feb 2004  0.29875209  0.25719228
#> Mar 2004 -1.31329880 -0.69076345
#> Apr 2004  2.48926902  1.54309187
#> May 2004  1.91175656  0.91165371
#> Jun 2004  2.04130891  1.23352769
#> Jul 2004 -0.29759583 -0.05351566
#> Aug 2004 -1.22025180 -0.56162404
#> Sep 2004 -0.95856611 -0.54609232
#> Oct 2004  0.20717781  0.29095534
#> Nov 2004 -0.70462371 -1.15899926
#> Dec 2004  2.26007518  1.40841386
#> Jan 2005  2.95319223  1.69957104
#> Feb 2005  0.89317411  0.77051186
#> Mar 2005 -1.54622760 -0.71981400
#> 
#> $coefficients
#>        Quotes   TV.advert
#> x1  0.4358443  0.18606264
#> x2  0.1361999  0.08192973
#> h1  0.2311472 -0.09036149
#> h2 -1.2318730 -0.60953989
#> h3 -0.1129694 -0.08956715
#> h4  0.4033300  0.15098381
#> h5 -0.4866789 -0.25350343
#> 
#> $loocv
#> [1] 1.924173
#> 
#> $weighted_loocv
#> [1] 3.601775e-17
#> 
#> $loocv_per_series
#>    Quotes TV.advert 
#>  2.601976  1.246370 
#> 
#> attr(,"class")
#> [1] "mtsforecast"
print(ahead::loocvridge2f(fpp::usconsumption))
#> $mean
#>         consumption    income
#> 2011 Q1   0.7373376 0.9197108
#> 2011 Q2   0.7523869 0.7436787
#> 2011 Q3   0.7537811 0.8142475
#> 2011 Q4   0.7563826 0.7913050
#> 2012 Q1   0.7568367 0.8009631
#> 
#> $lower
#>         consumption     income
#> 2011 Q1  -0.5227315 -0.7844227
#> 2011 Q2  -0.5076823 -0.9604548
#> 2011 Q3  -0.5062880 -0.8898860
#> 2011 Q4  -0.5036865 -0.9128285
#> 2012 Q1  -0.5032325 -0.9031703
#> 
#> $upper
#>         consumption   income
#> 2011 Q1    1.997407 2.623844
#> 2011 Q2    2.012456 2.447812
#> 2011 Q3    2.013850 2.518381
#> 2011 Q4    2.016452 2.495438
#> 2012 Q1    2.016906 2.505097
#> 
#> $sims
#> NULL
#> 
#> $x
#>         consumption       income
#> 1970 Q1  0.61227692  0.496540045
#> 1970 Q2  0.45492979  1.736459591
#> 1970 Q3  0.87467302  1.344880981
#> 1970 Q4 -0.27251439 -0.328145953
#> 1971 Q1  1.89218699  1.965432327
#> 1971 Q2  0.91337819  1.490757133
#> 1971 Q3  0.79285790  0.442927733
#> 1971 Q4  1.64999566  1.050230993
#> 1972 Q1  1.32724825  0.629713564
#> 1972 Q2  1.88990506  0.934207242
#> 1972 Q3  1.53272416  2.024456496
#> 1972 Q4  2.31705777  3.901784825
#> 1973 Q1  1.81385569  1.129195652
#> 1973 Q2 -0.05055772  0.981487307
#> 1973 Q3  0.35966722  0.503626895
#> 1973 Q4 -0.29331546  1.211474571
#> 1974 Q1 -0.87877094 -1.546943855
#> 1974 Q2  0.34672003 -0.860297695
#> 1974 Q3  0.41195356  0.310671244
#> 1974 Q4 -1.47820468 -0.452458059
#> 1975 Q1  0.83735987 -0.393048299
#> 1975 Q2  1.65397369  4.564235521
#> 1975 Q3  1.41431884 -1.359432036
#> 1975 Q4  1.05310993  1.006575406
#> 1976 Q1  1.97774749  1.462092078
#> 1976 Q2  0.91507218  0.819999864
#> 1976 Q3  1.05074607  0.896376224
#> 1976 Q4  1.29519619  0.715996597
#> 1977 Q1  1.13545889  0.082385052
#> 1977 Q2  0.55153240  1.131054305
#> 1977 Q3  0.95015960  1.490114704
#> 1977 Q4  1.49616150  1.995499552
#> 1978 Q1  0.58229978  0.614904129
#> 1978 Q2  2.11467168  1.141343133
#> 1978 Q3  0.41869886  0.808026906
#> 1978 Q4  0.80276430  0.782540078
#> 1979 Q1  0.50412878  1.087249832
#> 1979 Q2 -0.05855113 -0.700631545
#> 1979 Q3  0.97755597  0.568900915
#> 1979 Q4  0.26591209  0.797134552
#> 1980 Q1 -0.17368425  0.279831351
#> 1980 Q2 -2.29656300 -1.409415369
#> 1980 Q3  1.06691983  1.024445050
#> 1980 Q4  1.32441742  2.043130338
#> 1981 Q1  0.54583283 -0.229300048
#> 1981 Q2  0.00000000  0.013772840
#> 1981 Q3  0.40482184  2.193050433
#> 1981 Q4 -0.75874883  0.201893379
#> 1982 Q1  0.64399814  0.116463988
#> 1982 Q2  0.35685950  0.698172650
#> 1982 Q3  0.76412375  0.436938581
#> 1982 Q4  1.80788661  0.351269015
#> 1983 Q1  0.97593734  0.784231599
#> 1983 Q2  1.96559809  0.726007440
#> 1983 Q3  1.75134970  1.489909944
#> 1983 Q4  1.57374005  2.042140999
#> 1984 Q1  0.85322727  2.163513516
#> 1984 Q2  1.42002574  1.707008683
#> 1984 Q3  0.76950200  1.545359737
#> 1984 Q4  1.30747803  0.946974902
#> 1985 Q1  1.68128155 -0.246259445
#> 1985 Q2  0.90791081  1.966841071
#> 1985 Q3  1.88085044 -0.620830801
#> 1985 Q4  0.21986403  1.030018787
#> 1986 Q1  0.83153359  1.183355120
#> 1986 Q2  1.05966370  1.128116143
#> 1986 Q3  1.73244172  0.523753063
#> 1986 Q4  0.60006243  0.059897054
#> 1987 Q1 -0.15228800  0.611885129
#> 1987 Q2  1.32935729 -1.090188910
#> 1987 Q3  1.11041685  1.772266694
#> 1987 Q4  0.24012547  1.410476141
#> 1988 Q1  1.65692852  1.248862406
#> 1988 Q2  0.72306031  0.934536046
#> 1988 Q3  0.78681412  0.766858588
#> 1988 Q4  1.17068014  0.896061766
#> 1989 Q1  0.36522624  1.130999186
#> 1989 Q2  0.44694325 -0.411533933
#> 1989 Q3  1.03134287  0.645197097
#> 1989 Q4  0.48794531  0.762948357
#> 1990 Q1  0.78786794  0.750360325
#> 1990 Q2  0.32958888  0.655150176
#> 1990 Q3  0.37909401  0.072718052
#> 1990 Q4 -0.78228237 -0.676795704
#> 1991 Q1 -0.28358087  0.307585359
#> 1991 Q2  0.75819378  0.755588074
#> 1991 Q3  0.38256742  0.211957150
#> 1991 Q4 -0.04493204  0.649885878
#> 1992 Q1  1.70442848  1.497981581
#> 1992 Q2  0.59103346  0.763512573
#> 1992 Q3  1.09931218  0.506330196
#> 1992 Q4  1.21261583  1.412562187
#> 1993 Q1  0.40511535 -1.477541558
#> 1993 Q2  0.95540152  1.543168336
#> 1993 Q3  1.07908089  0.212590891
#> 1993 Q4  0.88609934  1.463027070
#> 1994 Q1  1.10781585 -0.400479849
#> 1994 Q2  0.73801073  1.691849180
#> 1994 Q3  0.79832641  0.732048430
#> 1994 Q4  0.98235581  1.317421015
#> 1995 Q1  0.11670364  0.646210704
#> 1995 Q2  0.81643944  0.009071803
#> 1995 Q3  0.89012327  0.745603285
#> 1995 Q4  0.70025668  0.591013847
#> 1996 Q1  0.90999511  1.059530012
#> 1996 Q2  1.12517415  1.068869375
#> 1996 Q3  0.59749105  0.847813936
#> 1996 Q4  0.81104981  0.547321447
#> 1997 Q1  1.00231479  0.890373141
#> 1997 Q2  0.40370845  0.748012589
#> 1997 Q3  1.68561876  1.138248868
#> 1997 Q4  1.13779625  1.379965421
#> 1998 Q1  0.98935016  2.248544314
#> 1998 Q2  1.70668759  1.417721496
#> 1998 Q3  1.31690105  1.045312610
#> 1998 Q4  1.52238359  0.730112366
#> 1999 Q1  0.98149855  0.675449171
#> 1999 Q2  1.56147049  0.228461748
#> 1999 Q3  1.19479035  0.644917550
#> 1999 Q4  1.40026421  1.555650840
#> 2000 Q1  1.50504064  2.079917327
#> 2000 Q2  0.93588274  1.019653998
#> 2000 Q3  0.97432184  1.055609188
#> 2000 Q4  0.88064976  0.149548646
#> 2001 Q1  0.39868539  0.748024500
#> 2001 Q2  0.37651229 -0.274110069
#> 2001 Q3  0.43918859  2.514669298
#> 2001 Q4  1.55369100 -1.173044414
#> 2002 Q1  0.34382689  2.659477606
#> 2002 Q2  0.50665404  0.549165469
#> 2002 Q3  0.67571194 -0.343744371
#> 2002 Q4  0.35472465  0.236234260
#> 2003 Q1  0.50387273  0.367138124
#> 2003 Q2  0.98555573  1.501775306
#> 2003 Q3  1.33766670  1.390030780
#> 2003 Q4  0.54254673  0.573803177
#> 2004 Q1  0.88074795  0.443164179
#> 2004 Q2  0.44397949  0.983686803
#> 2004 Q3  0.87037870  0.666082163
#> 2004 Q4  1.07395152  1.388903698
#> 2005 Q1  0.79393888 -1.225522766
#> 2005 Q2  0.98477889  0.701673549
#> 2005 Q3  0.75627802  0.594778066
#> 2005 Q4  0.24819787  0.545328307
#> 2006 Q1  1.01902713  1.855678610
#> 2006 Q2  0.60048219  0.882341774
#> 2006 Q3  0.59799998  0.479144681
#> 2006 Q4  0.92584113  1.302357376
#> 2007 Q1  0.55424125  0.451389832
#> 2007 Q2  0.38257957  0.149334878
#> 2007 Q3  0.43929546  0.392083863
#> 2007 Q4  0.29465872  0.548559897
#> 2008 Q1 -0.25266521  1.430770981
#> 2008 Q2 -0.03553182  1.973964634
#> 2008 Q3 -0.97177447 -2.308604067
#> 2008 Q4 -1.31350400 -0.057706851
#> 2009 Q1 -0.38748400 -0.969060881
#> 2009 Q2 -0.47008302  0.063290188
#> 2009 Q3  0.57400096 -1.392556217
#> 2009 Q4  0.10932885 -0.144713401
#> 2010 Q1  0.67101795  1.187165135
#> 2010 Q2  0.71771819  1.354354721
#> 2010 Q3  0.65314326  0.561169813
#> 2010 Q4  0.87535215  0.371057940
#> 
#> $level
#> [1] 95
#> 
#> $method
#> [1] "ridge2"
#> 
#> $fitted
#>         consumption      income
#> 1970 Q1  0.68809125  0.79489953
#> 1970 Q2  0.65546116  0.26319874
#> 1970 Q3  0.82292378  0.69828401
#> 1970 Q4  0.37978311  0.46611712
#> 1971 Q1  1.29051752  1.13343541
#> 1971 Q2  0.84354448  0.67665061
#> 1971 Q3  0.72607367  0.87586180
#> 1971 Q4  1.16939724  1.11654996
#> 1972 Q1  0.94948176  1.05720293
#> 1972 Q2  1.23499030  1.22802777
#> 1972 Q3  1.12227108  0.94146975
#> 1972 Q4  1.50943754  0.86913997
#> 1973 Q1  1.25032684  1.16782487
#> 1973 Q2  0.50363734  0.14499985
#> 1973 Q3  0.64238988  0.60877493
#> 1973 Q4  0.41677677 -0.10904347
#> 1974 Q1  0.05503699  0.48654750
#> 1974 Q2  0.44777155  0.88381381
#> 1974 Q3  0.61638586  0.73063180
#> 1974 Q4  0.14550869 -0.42070118
#> 1975 Q1  0.62972071  1.01500101
#> 1975 Q2  1.24613868  0.16693656
#> 1975 Q3  0.64891736  1.38885508
#> 1975 Q4  0.89658997  0.90384617
#> 1976 Q1  1.34594429  1.19515246
#> 1976 Q2  0.83251625  0.86791347
#> 1976 Q3  0.89820881  0.91004964
#> 1976 Q4  0.95688241  1.03118481
#> 1977 Q1  0.76487878  1.06695903
#> 1977 Q2  0.67933871  0.53818545
#> 1977 Q3  0.85917713  0.70335074
#> 1977 Q4  1.10588920  0.92497116
#> 1978 Q1  0.70175251  0.73539017
#> 1978 Q2  1.36856806  1.28283027
#> 1978 Q3  0.65217848  0.54741094
#> 1978 Q4  0.77748346  0.83139234
#> 1979 Q1  0.65792136  0.51891977
#> 1979 Q2  0.36758441  0.69641346
#> 1979 Q3  0.80136629  0.93084992
#> 1979 Q4  0.60862597  0.43842046
#> 1980 Q1  0.50797842  0.28745758
#> 1980 Q2 -0.17259804 -0.65573193
#> 1980 Q3  0.90302040  0.90749177
#> 1980 Q4  1.03416547  0.78515496
#> 1981 Q1  0.57832124  0.87311387
#> 1981 Q2  0.48995227  0.53460642
#> 1981 Q3  0.64713283  0.07218524
#> 1981 Q4  0.35383648 -0.12175018
#> 1982 Q1  0.64713907  0.86281576
#> 1982 Q2  0.64208173  0.53778152
#> 1982 Q3  0.71816428  0.86508243
#> 1982 Q4  1.07038717  1.28776915
#> 1983 Q1  0.84991822  0.89637800
#> 1983 Q2  1.21650000  1.29016857
#> 1983 Q3  1.23238260  1.11494590
#> 1983 Q4  1.14022708  0.96499972
#> 1984 Q1  0.83709417  0.40505418
#> 1984 Q2  1.06528295  0.96804893
#> 1984 Q3  0.78387727  0.55453315
#> 1984 Q4  1.01434329  0.99965613
#> 1985 Q1  0.88533777  1.33247468
#> 1985 Q2  0.85476597  0.51116309
#> 1985 Q3  0.87640396  1.46890891
#> 1985 Q4  0.57863298  0.32833878
#> 1986 Q1  0.79997089  0.72203527
#> 1986 Q2  0.89677186  0.89837152
#> 1986 Q3  1.08081661  1.23133812
#> 1986 Q4  0.62895388  0.85320144
#> 1987 Q1  0.50051732  0.19092910
#> 1987 Q2  0.66239993  1.31541411
#> 1987 Q3  0.93539771  0.72299035
#> 1987 Q4  0.55733989  0.21884438
#> 1988 Q1  1.19143369  1.09759506
#> 1988 Q2  0.74673357  0.72836848
#> 1988 Q3  0.76993059  0.82686581
#> 1988 Q4  0.95020728  0.95449038
#> 1989 Q1  0.61359729  0.40239852
#> 1989 Q2  0.53033471  0.85942913
#> 1989 Q3  0.83943532  0.93976341
#> 1989 Q4  0.67548950  0.61316430
#> 1990 Q1  0.76999352  0.82860674
#> 1990 Q2  0.63724189  0.53176178
#> 1990 Q3  0.57569488  0.76177781
#> 1990 Q4  0.22674665  0.21566595
#> 1991 Q1  0.48415602  0.19372492
#> 1991 Q2  0.75659096  0.81438711
#> 1991 Q3  0.59444816  0.74301528
#> 1991 Q4  0.52895222  0.25758711
#> 1992 Q1  1.20878826  1.09817602
#> 1992 Q2  0.70535979  0.68892115
#> 1992 Q3  0.83381326  0.98800549
#> 1992 Q4  0.96864849  0.91863196
#> 1993 Q1  0.38297473  0.99689856
#> 1993 Q2  0.86291738  0.68912200
#> 1993 Q3  0.76759952  1.02524517
#> 1993 Q4  0.83114789  0.66641993
#> 1994 Q1  0.69596887  1.12570299
#> 1994 Q2  0.77464601  0.48215112
#> 1994 Q3  0.76982624  0.83554953
#> 1994 Q4  0.86796266  0.78513230
#> 1995 Q1  0.57611260  0.37786025
#> 1995 Q2  0.67618901  0.94826938
#> 1995 Q3  0.80815756  0.86909675
#> 1995 Q4  0.72209101  0.81687521
#> 1996 Q1  0.82983352  0.82054450
#> 1996 Q2  0.93093160  0.92484263
#> 1996 Q3  0.70121272  0.66610931
#> 1996 Q4  0.74400709  0.86811297
#> 1997 Q1  0.87421898  0.89365352
#> 1997 Q2  0.65211523  0.55599441
#> 1997 Q3  1.20317093  1.11656169
#> 1997 Q4  0.93588425  0.87582400
#> 1998 Q1  0.89743581  0.47420711
#> 1998 Q2  1.21195162  1.10407322
#> 1998 Q3  1.02710226  0.99276617
#> 1998 Q4  1.04731102  1.11726451
#> 1999 Q1  0.82721116  0.91563639
#> 1999 Q2  0.94775673  1.21129875
#> 1999 Q3  0.90211025  1.00333253
#> 1999 Q4  1.05570353  0.98912555
#> 2000 Q1  1.11206966  0.90271828
#> 2000 Q2  0.83971426  0.85271451
#> 2000 Q3  0.85709374  0.86818938
#> 2000 Q4  0.71019326  0.95394638
#> 2001 Q1  0.65065696  0.55229020
#> 2001 Q2  0.53049097  0.81097349
#> 2001 Q3  0.67090860 -0.01220709
#> 2001 Q4  0.70749672  1.41834714
#> 2002 Q1  0.63445186 -0.13000454
#> 2002 Q2  0.67649019  0.70411238
#> 2002 Q3  0.59588921  0.94233710
#> 2002 Q4  0.59276993  0.71735589
#> 2003 Q1  0.64452866  0.76970484
#> 2003 Q2  0.87457044  0.72487867
#> 2003 Q3  1.02870363  0.97799320
#> 2003 Q4  0.68728175  0.72170905
#> 2004 Q1  0.74794300  0.91145251
#> 2004 Q2  0.64696810  0.50858854
#> 2004 Q3  0.78241911  0.87392094
#> 2004 Q4  0.90897192  0.82682357
#> 2005 Q1  0.51197052  1.11799468
#> 2005 Q2  0.83445894  0.91278972
#> 2005 Q3  0.73649486  0.83903736
#> 2005 Q4  0.62147000  0.50771877
#> 2006 Q1  0.89888394  0.62888651
#> 2006 Q2  0.69961465  0.65702120
#> 2006 Q3  0.68230866  0.79163250
#> 2006 Q4  0.84348632  0.74955405
#> 2007 Q1  0.66786937  0.77791614
#> 2007 Q2  0.58640512  0.75209178
#> 2007 Q3  0.63602819  0.71769627
#> 2007 Q4  0.63471962  0.54088801
#> 2008 Q1  0.41290935 -0.15081399
#> 2008 Q2  0.45352256 -0.17052058
#> 2008 Q3 -0.06593488  0.55918532
#> 2008 Q4  0.21141702 -0.44552200
#> 2009 Q1  0.25136176  0.60197042
#> 2009 Q2  0.41647688  0.14764767
#> 2009 Q3  0.43585888  1.05304277
#> 2009 Q4  0.48072673  0.68392990
#> 2010 Q1  0.73177799  0.60517323
#> 2010 Q2  0.75640721  0.58206382
#> 2010 Q3  0.70654972  0.80210172
#> 
#> $residuals
#>           consumption      income
#> 1970 Q1 -0.2331614550  0.94156006
#> 1970 Q2  0.2192118592  1.08168224
#> 1970 Q3 -1.0954381614 -1.02642996
#> 1970 Q4  1.5124038826  1.49931520
#> 1971 Q1 -0.3771393349  0.35732173
#> 1971 Q2 -0.0506865853 -0.23372287
#> 1971 Q3  0.9239219899  0.17436919
#> 1971 Q4  0.1578510125 -0.48683640
#> 1972 Q1  0.9404233021 -0.12299569
#> 1972 Q2  0.2977338655  0.79642872
#> 1972 Q3  1.1947866905  2.96031507
#> 1972 Q4  0.3044181569  0.26005568
#> 1973 Q1 -1.3008845581 -0.18633757
#> 1973 Q2 -0.1439701144  0.35862704
#> 1973 Q3 -0.9357053427  0.60269964
#> 1973 Q4 -1.2955477147 -1.43790038
#> 1974 Q1  0.2916830364 -1.34684520
#> 1974 Q2 -0.0358179935 -0.57314256
#> 1974 Q3 -2.0945905363 -1.18308986
#> 1974 Q4  0.6918511853  0.02765288
#> 1975 Q1  1.0242529803  3.54923451
#> 1975 Q2  0.1681801533 -1.52636860
#> 1975 Q3  0.4041925684 -0.38227968
#> 1975 Q4  1.0811575182  0.55824591
#> 1976 Q1 -0.4308721073 -0.37515259
#> 1976 Q2  0.2182298131  0.02846275
#> 1976 Q3  0.3969873869 -0.19405305
#> 1976 Q4  0.1785764797 -0.94879976
#> 1977 Q1 -0.2133463786  0.06409528
#> 1977 Q2  0.2708208867  0.95192925
#> 1977 Q3  0.6369843631  1.29214881
#> 1977 Q4 -0.5235894228 -0.31006703
#> 1978 Q1  1.4129191679  0.40595296
#> 1978 Q2 -0.9498691939 -0.47480336
#> 1978 Q3  0.1505858209  0.23512913
#> 1978 Q4 -0.2733546751  0.25585749
#> 1979 Q1 -0.7164724886 -1.21955132
#> 1979 Q2  0.6099715558 -0.12751255
#> 1979 Q3 -0.5354542028 -0.13371536
#> 1979 Q4 -0.7823102270 -0.15858911
#> 1980 Q1 -2.8045414131 -1.69687295
#> 1980 Q2  1.2395178705  1.68017698
#> 1980 Q3  0.4213970217  1.13563857
#> 1980 Q4 -0.4883326449 -1.01445500
#> 1981 Q1 -0.5783212443 -0.85934103
#> 1981 Q2 -0.0851304264  1.65844402
#> 1981 Q3 -1.4058816648  0.12970813
#> 1981 Q4  0.2901616584  0.23821417
#> 1982 Q1 -0.2902795770 -0.16464311
#> 1982 Q2  0.1220420227 -0.10084294
#> 1982 Q3  1.0897223303 -0.51381341
#> 1982 Q4 -0.0944498302 -0.50353755
#> 1983 Q1  1.1156798734 -0.17037056
#> 1983 Q2  0.5348497048  0.19974137
#> 1983 Q3  0.3413574446  0.92719510
#> 1983 Q4 -0.2869998070  1.19851380
#> 1984 Q1  0.5829315611  1.30195451
#> 1984 Q2 -0.2957809447  0.57731081
#> 1984 Q3  0.5236007542  0.39244175
#> 1984 Q4  0.6669382649 -1.24591558
#> 1985 Q1  0.0225730325  0.63436639
#> 1985 Q2  1.0260844738 -1.13199389
#> 1985 Q3 -0.6565399212 -0.43889013
#> 1985 Q4  0.2529006085  0.85501634
#> 1986 Q1  0.2596928119  0.40608088
#> 1986 Q2  0.8356698597 -0.37461846
#> 1986 Q3 -0.4807541752 -1.17144107
#> 1986 Q4 -0.7812418809 -0.24131632
#> 1987 Q1  0.8288399690 -1.28111801
#> 1987 Q2  0.4480169178  0.45685259
#> 1987 Q3 -0.6952722392  0.68748579
#> 1987 Q4  1.0995886217  1.03001803
#> 1988 Q1 -0.4683733788 -0.16305901
#> 1988 Q2  0.0400805474  0.03849011
#> 1988 Q3  0.4007495488  0.06919595
#> 1988 Q4 -0.5849810399  0.17650881
#> 1989 Q1 -0.1666540429 -0.81393245
#> 1989 Q2  0.5010081587 -0.21423203
#> 1989 Q3 -0.3514900073 -0.17681505
#> 1989 Q4  0.1123784390  0.13719603
#> 1990 Q1 -0.4404046425 -0.17345656
#> 1990 Q2 -0.2581478848 -0.45904373
#> 1990 Q3 -1.3579772508 -1.43857351
#> 1990 Q4 -0.5103275184  0.09191941
#> 1991 Q1  0.2740377581  0.56186316
#> 1991 Q2 -0.3740235474 -0.60242996
#> 1991 Q3 -0.6393802004 -0.09312940
#> 1991 Q4  1.1754762572  1.24039448
#> 1992 Q1 -0.6177548032 -0.33466345
#> 1992 Q2  0.3939523863 -0.18259096
#> 1992 Q3  0.3788025719  0.42455670
#> 1992 Q4 -0.5635331385 -2.39617352
#> 1993 Q1  0.5724267884  0.54626978
#> 1993 Q2  0.2161635095 -0.47653111
#> 1993 Q3  0.1184998238  0.43778190
#> 1993 Q4  0.2766679524 -1.06689977
#> 1994 Q1  0.0420418617  0.56614619
#> 1994 Q2  0.0236804018  0.24989731
#> 1994 Q3  0.2125295749  0.48187149
#> 1994 Q4 -0.7512590281 -0.13892159
#> 1995 Q1  0.2403268368 -0.36878845
#> 1995 Q2  0.2139342604 -0.20266609
#> 1995 Q3 -0.1079008757 -0.27808290
#> 1995 Q4  0.1879040963  0.24265480
#> 1996 Q1  0.2953406378  0.24832487
#> 1996 Q2 -0.3334405505 -0.07702869
#> 1996 Q3  0.1098370901 -0.11878787
#> 1996 Q4  0.2583077034  0.02226017
#> 1997 Q1 -0.4705105337 -0.14564093
#> 1997 Q2  1.0335035262  0.58225446
#> 1997 Q3 -0.0653746871  0.26340373
#> 1997 Q4  0.0534659027  1.37272032
#> 1998 Q1  0.8092517747  0.94351439
#> 1998 Q2  0.1049494266 -0.05876061
#> 1998 Q3  0.4952813338 -0.26265381
#> 1998 Q4 -0.0658124656 -0.44181534
#> 1999 Q1  0.7342593334 -0.68717464
#> 1999 Q2  0.2470336223 -0.56638120
#> 1999 Q3  0.4981539600  0.55231831
#> 1999 Q4  0.4493371036  1.09079178
#> 2000 Q1 -0.1761869235  0.11693572
#> 2000 Q2  0.1346075806  0.20289467
#> 2000 Q3  0.0235560235 -0.71864073
#> 2000 Q4 -0.3115078723 -0.20592188
#> 2001 Q1 -0.2741446687 -0.82640027
#> 2001 Q2 -0.0913023803  1.70369581
#> 2001 Q3  0.8827823993 -1.16083732
#> 2001 Q4 -0.3636698347  1.24113047
#> 2002 Q1 -0.1277978244  0.67917001
#> 2002 Q2 -0.0007782507 -1.04785675
#> 2002 Q3 -0.2411645608 -0.70610284
#> 2002 Q4 -0.0888971987 -0.35021777
#> 2003 Q1  0.3410270697  0.73207046
#> 2003 Q2  0.4630962693  0.66515211
#> 2003 Q3 -0.4861568985 -0.40419002
#> 2003 Q4  0.1934661935 -0.27854487
#> 2004 Q1 -0.3039635158  0.07223429
#> 2004 Q2  0.2234105933  0.15749362
#> 2004 Q3  0.2915324118  0.51498276
#> 2004 Q4 -0.1150330390 -2.05234634
#> 2005 Q1  0.4728083725 -0.41632113
#> 2005 Q2 -0.0781809229 -0.31801165
#> 2005 Q3 -0.4882969879 -0.29370905
#> 2005 Q4  0.3975571281  1.34795984
#> 2006 Q1 -0.2984017519  0.25345526
#> 2006 Q2 -0.1016146655 -0.17787652
#> 2006 Q3  0.2435324690  0.51072488
#> 2006 Q4 -0.2892450783 -0.29816422
#> 2007 Q1 -0.2852898034 -0.62858127
#> 2007 Q2 -0.1471096634 -0.36000792
#> 2007 Q3 -0.3413694729 -0.16913638
#> 2007 Q4 -0.8873848342  0.88988298
#> 2008 Q1 -0.4484411727  2.12477862
#> 2008 Q2 -1.4252970242 -2.13808348
#> 2008 Q3 -1.2475691203 -0.61689217
#> 2008 Q4 -0.5989010146 -0.52353888
#> 2009 Q1 -0.7214447804 -0.53868023
#> 2009 Q2  0.1575240776 -1.54020389
#> 2009 Q3 -0.3265300272 -1.19775618
#> 2009 Q4  0.1902912192  0.50323524
#> 2010 Q1 -0.0140597983  0.74918149
#> 2010 Q2 -0.1032639558 -0.02089401
#> 2010 Q3  0.1688024258 -0.43104378
#> 
#> $coefficients
#>     consumption      income
#> x1  0.102077302  0.12949577
#> x2  0.006601867 -0.02957051
#> h1  0.092473613  0.04771835
#> h2 -0.053810426 -0.20499353
#> h3 -0.014413797 -0.05815312
#> h4  0.099574873  0.03766025
#> h5 -0.136362528  0.05021583
#> 
#> $loocv
#> [1] 0.6183515
#> 
#> $weighted_loocv
#> [1] -2.493598e-17
#> 
#> $loocv_per_series
#> consumption      income 
#>   0.4371489   0.7995541 
#> 
#> attr(,"class")
#> [1] "mtsforecast"

#foo <- function(xx) ahead::loocvridge2f(fpp::insurance, lambda_1=10^xx[1], lambda_2=10^xx[2])
#(opt <- stats::nlminb(objective=foo, lower=c(-10,-10), upper=c(10,10), start=c(0, 0)))
#print(ahead::loocvridge2f(fpp::insurance, lambda_1=10^opt$par[1], lambda_2=10^opt$par[2]))